feat: mapper

+++ src/main/java/br/com/fiap/study_apir/controller/ProdutoController.java
@@ -1,7 +1,6 @@
 package br.com.fiap.study_apir.controller;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -17,8 +16,11 @@ import org.springframework.web.bind.annotation.RestController;
 
 import br.com.fiap.study_apir.dto.ProdutoCreateRequest;
 import br.com.fiap.study_apir.dto.ProdutoMapper;
+import br.com.fiap.study_apir.dto.ProdutoResponse;
+import br.com.fiap.study_apir.dto.ProdutoUpdateRequest;
 import br.com.fiap.study_apir.model.Produto;
 import br.com.fiap.study_apir.service.ProdutoService;
+import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("api/${api.version}/produtos")
@@ -31,36 +33,39 @@ public class ProdutoController {
     private ProdutoMapper produtoMapper;
 
     @PostMapping
-    public ResponseEntity<Produto> create(@RequestBody ProdutoCreateRequest dtoRequest) {                 
-        
-        return ResponseEntity.status(HttpStatus.CREATED).body(
-                        service.createOrUpdate(
-                            produtoMapper.toModel(dtoRequest)));
+    public ResponseEntity<ProdutoResponse> create(@Valid @RequestBody ProdutoCreateRequest dtoRequest) {                         
+        return ResponseEntity.status(HttpStatus.CREATED).body(produtoMapper.toDto(
+                service.createOrUpdate(
+                    produtoMapper.toModel(dtoRequest)
+                )
+            ));
     }
 
     @GetMapping("/{id}")    
-    public ResponseEntity<Produto> findById(@PathVariable Long id) { 
+    public ResponseEntity<ProdutoResponse> findById(@PathVariable Long id) { 
         return service
                 .findById(id)
+                .map(produto -> produtoMapper.toDto(produto))                
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());        
     }
 
     @GetMapping    
-    public ResponseEntity<List<Produto>> findAll() {        
-        return ResponseEntity.ok(service.findAll());
+    public ResponseEntity<List<ProdutoResponse>> findAll() {        
+        return ResponseEntity.ok(
+                service.findAll().stream()
+                .map(produto -> produtoMapper.toDto(produto))
+                .toList());
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Produto> update(@PathVariable Long id, 
-                                @RequestBody Produto produto) {
+    public ResponseEntity<ProdutoResponse> update(@PathVariable Long id, 
+                                @RequestBody ProdutoUpdateRequest dtoRequest) {
 
-        Optional<Produto> optProduto = service.findById(id);
-
-        if (optProduto.isPresent()) {
-            produto.setId(id);
-            Produto produtoAlterado = service.createOrUpdate(produto);
-            return ResponseEntity.ok(produtoAlterado);
+        if (service.findById(id).isPresent()) {
+            Produto produto = produtoMapper.toModel(id, dtoRequest);
+            return ResponseEntity.ok(produtoMapper.toDto(
+                            service.createOrUpdate(produto)));
         } else {
             return ResponseEntity.notFound().build();
         }     
@@ -68,8 +73,11 @@ public class ProdutoController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteById(@PathVariable Long id) { 
-        service.deleteById(id);
-        return ResponseEntity.noContent().build();
- 
+        if (service.findById(id).isPresent()) {
+            service.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }         
     }
 }
